+++ catroidTest/src/org/catrobat/catroid/test/content/actions/SetVariableActionTest.java
@@ -99,7 +99,7 @@ public class SetVariableActionTest extends AndroidTestCase {
 		userVariable = ProjectManager.getInstance().getCurrentProject().getUserVariables()
 				.getUserVariable(TEST_USERVARIABLE, null);
 
-		assertEquals("Variable not changed", CHANGE_VARIABLE_VALUE, userVariable.getValue().intValue());
+		assertEquals("Variable not changed", CHANGE_VARIABLE_VALUE, ((Double) userVariable.getValue()).intValue());
 		ProjectManager.getInstance().getCurrentProject().getUserVariables().deleteUserVariableByName(TEST_USERVARIABLE);
 
 	}
@@ -139,7 +139,7 @@ public class SetVariableActionTest extends AndroidTestCase {
 		userVariable = ProjectManager.getInstance().getCurrentProject().getUserVariables()
 				.getUserVariable(TEST_USERVARIABLE, null);
 
-		assertEquals("Variable changed, but should not!", 0, userVariable.getValue().intValue());
+		assertEquals("Variable changed, but should not!", 0, ((Double) userVariable.getValue()).intValue());
 		ProjectManager.getInstance().getCurrentProject().getUserVariables().deleteUserVariableByName(TEST_USERVARIABLE);
 
 	}
